Fix 'Attempt to overwrite cell' issue
If the number of columns are different among the dataset when
`auto_unite=True`, the number of columns of the first dataset was used
and that cause 'Attempt to overwrite cell' issue.

I change it to use the number of longest dataset columns instead of the number
of first dataset columns.

And also, some related issue were fixed (the position of Avg and Std or
so on).